chore: Refactor linting into `lib.rs`
See `lib.rs` for detailed explanation.

The idea here is to centralise lint requirements as much as possible.
Avoiding unnecessary repetition. Perhaps it could be thought of as a bit
like the "cascade" in CSS. There's a base "pendantic" requirement for as
much lint feedback as possible. Then crate-wide exceptions are made.
Then file-wide exceptions are added. And finally function-specific
exceptions are made on a case by case bases.

Note that by default lints will just produce warnings. Only in CI will
warnings be converted to failures.