chore: Remove linting definitions from `.vscode`
Lints are now formalised in the source code itself. But it's still good
to give VSCode users the headstart of defaulting to Clipp over `cargo
check`.

I added a `etc/NOTES.md` to keep a record of the pendantic VSCode
settings. I personally use something like as my default Neovim config.
But having a `etc/NOTES.md` might not be something you want.