chore: Remove `#[inline]`s
Copied comment from `lib.rs` linting:

Although performance is of course important for this application,
it is not currently such that it would benefit from explicit
inline suggestions. besides, not specifying `[inline]` doesn't
mean that a function won't be inlined. And if performance does
start to become a problem, there are other avenues to explore
before deciding on which functions would benefit from explicit
inlining.