feat: Basic structure for custom request logging
All HTTP requests are now logged with useful data. And other traces,
debugs, infos, etc, that are logged during the lifetime of a HTTP request
are also prepened with the that same HTTP context data.

Also an example of adding custom request data and conditionally acting
on that data is included. Namely, adding request duration and logging a
warning when the duration is over a certain amount of time.