refact: move `App` into a seperate function
This is my attempt of introducing a testing framework for testing
dynamic routing to git repositories. My initial idea is to use the same
`App` instance in the testing framework as I'm using in the production
server. However, the function signature is verbose and we can't use
lazy_static! to only initialize the App once in the testing framework,
as the function signature has dynamic dispatch.