feat(test): test basic routing to the law-html repository
By utilizing `test::TestRequest`, we can make Http requests to the test
instance of our Actix server. These tests should confirm that Http
requests are resolved correctly and are serving the current versions of
the documents. One current drawback is that the repositories are bare,
and it's difficult to reason about in the repository state.