feat: use `git2` for initializing archive/stele
Instead of relying on bash scripts, try using rust libgit2 bindings to
initialize the repo states. Benefit of using Rust libgit2 to initialize
repositories is that we're flexibile to update repositories during test
runs. We also don't need to additionally configure Windows test-runners
to install anything to be able to run bash scripts.