refact: fix `repositories` deserialization
Make `repositories` field a HashMap, instead of Vec. This update makes
the structure of `Repositories` type more readable, as the
struct should now directly map to how `repositories.json` fields are.

Actix still needs to create routes based on the longest to shortest
first. To resolve, add `Repositories` method that returns repositories
sorted by routes.

For coding ergonomics, it's simpler to have each Repository object have
a reference to it's key. To resolve, add a `name` field and use
`RepositoriesVisitor` for deserialization.