feat: add more modules for inserting history
The purpose of the submodule is to store any/all functions that work
with rdf. Currently the underlying implementation relies on `sophia` to
parse the graph and retrieve triples.

Add `namespaces.rs` module that stores rdf namespaces. Currently stores
the `oll` and `dcterms` ontology. I expected the `dcterms` to be
supported out-of-the-box, but `sophia` supports only the most frequently
used namespaces.

*Add sqlx chrono feature*

Will be necessary to work with dates in the database.

*Add statements module for inserts and queries in db*

The idea being that this module stores queries/insert
statements for interacting with both databases.

*Add chrono as dependency*

Necessary because chrono exports serde Serialize and Deserialize traits.
We then use these traits to map sqlx queries to structs.

*Add models mod for sqlx structs*

Created a centralized place for all database related structs.
This approach is convenient because the sqlx queries will then
serialize/deserialize to these structs using serde.

*Rename from upsert to create*

Much easier to reason about what the function is supposed to do.

*Wrap insert changes into an atomic transaction*