feat(history): add `rdf` submodule
The purpose of the submodule is to store any/all functions that work
with rdf. Currently the underlying implementation relies on `sophia` to
parse the graph and retrieve triples.

Add `namespaces.rs` module that stores rdf namespaces. Currently stores
the `oll` and `dcterms` ontology. I expected the `dcterms` to be
supported out-of-the-box, but `sophia` supports only the most frequently
used namespaces.