feat: further work on inserting changes into db
The idea being we iterate over all publications using libgit2, and then
for each publication we parse .rdf xml files into an rdf graph.
After the graph is loaded, we iterate over document and collection
versions from the graph and insert each into db.

*Add more sqlx structs for database*

*Wrap sophia graph with our StelaeGraph*

Goal is to move similar code for working with rdf graph into a separate wrapper struct.
This simplifies working with graphs as it hides away the rdf graph results complexity.

*Use composite keys instead of autoincrement synthetic keys*

*Split functionalities for working with rdf graph on*

Since the graph queries will vary on the type of information that we
need, the easiest is to create utility functions that return (s,p,o)
depending on the queries.