feat: implement the rest of insert changes.
This commit has the following significant changes:

*Bulk insert multiple document and collection change objects into db*

Idea being we create a query-builder and insert multiple values into an insert statement.

On Windows, sqlx throws an error when inserting 4000+ rows at once.
To resolve, chunk the rows to 2000 per bulk insert. A
dd `BATCH_SIZE` constant, indicating how many rows will get inserted at once. Set to 2000 per insert.

*Upgrade rustc compiler version*

*Add utility methods for working with the `sophia` graph*

RDF elements can be stored in a container, e.g.
<rdf:_1 resource="...">
<rdf:_2 resource="...">
To resolve, introduce a `Bag` struct with `items` helper method.
The idea being we iterate over all elements inside a container. This logic
is similar to rdflib implementation of `items` in Python.
To better understand this iteration logic, please see `rdflib/container.py` in rdflib Python.

*Add missing sqlx structs to map objects to database rows*

*Revoke same date publications*

Set a boolean revoked to True for publications that exist on same date.
Revoked publications aren't accounted for when querying for versions for a publication.

*Support for publication partial update*

Re-use the existing insert function, as most of the code is the same.
The only difference is that we skip parsing/loading publications which
are already inserted in the db. For publication directories which aren't
in the DB, we continue inserting changes.