feat(rdf): add `Bag` container struct for working with sequences
RDF elements can be stored in a container, e.g.
<rdf:_1 resource="...">
<rdf:_2 resource="...">

To resolve, introduce a `Bag` struct with `items` helper method. The
idea being we iterate over all elements inside a container. This logic
is similar to rdflib implementation of `items` in Python. To better
understand this iteration logic, please see
`rdflib/container.py` in rdflib Python.