refact(db): re-organize models with db `Manager` and `TxManager` traits
Two notable changes:

Remove `inserts.rs` and `queries.rs` modules and move existing methods
to it's own `models/*/manager.rs` module.
This change makes it really clear which model the method belongs to, and
hopefully increase codebase legibility. It should now be quite obvious
where in our codebase we're calling out to database operations.

Introduce `TxManager` trait, which holds all the methods that work with
a database transaction. Currently these methods are functions that
insert change objects in the database. Ideally we would want the
`Manager` and `TxManager` to be a single trait, but we would need to
know which Pool or Transaction object we use at compile time. This
seemed too tricky to figure out quickly, so instead opt to call out to
`TxManager` when managing transactions, and `Manager` when managing
regular connection calls (outside of a transaction).

Fix insert for multiple stele. Previously everything was getting applied
to a connection, instead of a transaction. This was because we used
`DatabaseConnection` struct everywhere. What we instead wanted was the
`DatabaseTransaction` struct, which runs the sqlx queries/inserts within
a transaction.

Add error logging to git2::walk method and fix the `find_blob` method to
not fail during the git repo walk.