refact: progress on centralizing errors
Progress on https://github.com/openlawlibrary/stelae/issues/36

Added `CliError` enum that maps any/all errors related to working with
stelae CLI.

Updated `run` to not return any errors, but instead to decide how to
stop the process. We expect to currently return error code 1 on any
found stelae errors, and instruct the users to inspect the local logs.

Still have issues with centralizing errors found during initializing
actix `App` instance. The reason this error handling is tricky is
because we're passing in the generic, opaque `App` type to both `app.rs`
and `routes.rs`. I tried initializing the `init_app` in `app.rs` outside
of the `HttpServer::new(..)`, but the problem was that `App` did not
implement `Clone` trait, so could not get it working quickly. This is
something we'd like to resolve in the future, so leave the process
exits uncentralized for now, and leave a comment in `app.rs`.