fix(db): add more indexes to optimize versions endpoint performance
For document or collection versions we want the query times to be really
fast. Since for both types of queries we do LEFT JOINS to retrieve
relevant versions, index the `publication_version` and
`publication_has_publication_versions` ids. Add a couple of additional
indexes for the `document_change` and `library_change` tables.