feat: add `commit_type` column to `data_repo_commits`
The idea being that we can't always reliably rely on
`auth_commit_timestamp` column to get the order of inserted commit
hashes. Mainly because it can sometimes happen that the timestamp date is the
same for all commits on a publication, thus making it impossible to
figure out which commit was inserted last.